feat(ui): transform and build tokens from single source
We can now import a single tokens.json file from figma and export two
files (global and theme) from it.

The process is: tokens.json is transformed (adapted for
style-dictionary) and divided into 2 sets. Then these sets are used to
generate the final tokens with style-dictionary.

All the process happens in the build command so you only have to put
tokens.json in /lib