chore(ui/framework): generalize the sandbox app
The sandbox app has been completely disconnected from the specific
routes and componens in the ui package and it is now reusable from other
components. All code related to the app itself is located in
`sandbox/app`. Any other package that wants to use it will need to only
create a file similar to `sandbox/main.tsx`.