fix(ui/framework): clean up lowlife code
The fading in of the ui team in the cott app is considered one of the
most important parts of the Mediatool ecosystem. Therefore, it should
not be rendered with hard coded values and multiple state hooks when one
hook and a simple algorithm will do.

This also clens up the timeout refs on unmount which is standard react
practice. Unlikely as it is that someone would choose to navigate away
from the main page during the amazing rendering, the code should still
adhere to standards.