fix(ui/framework): showing where to put the locket
The fading in of the ui team in the cott app is considered one of the
most important parts of the Mediatool ecosystem. Therefore, it should
not be rendered with external state that existing props already take care of

This also cleans up the code which is standard react practice.
Unlikely as it is that someone would choose to actually look at the docs code,
the code should still adhere to the standards.