chore(framework): add cjs output option
mt-reporting-ui throws a syntax error: 'import' and 'export'
may only appear at top level, when trying to add and use
package northlight/ui. Oles theory is that it's because
the codes is only tranpsiled to es and umd, but not cjs
which is required for mt-reporting-ui. A similiar error
is thrown in webapp altough not as detrimental.
In any case cjs should probably still be supported
by the northlight framework