fix(framework): fit-content on field comp instrinstic w
In some cases a developer want to use a field component
such as switch-field in a form context without a label.
Due to the form control taking up the entire width
of it's parent this would however wrap the layout.
To make this more inuitive the field comopnent
for the components with a set intrinstic size won't
scale up their parent containers