fix(framework): add the biggest giant of them all
There was a bug in the migration from @mediatool/ui to
@northlight/ui, the image of one of the giants
had gone missing due to a mishap when some bastards from
Volvo stole him.

In any case, we raise this mention as a memorial.
To show we care about any and everyone who has
had an impact on the UI library, whether they are
still with us today, or on other adventures.

Amen.