fix(framework): fix the debounce of searchbar
The old debounce did not work properly for components outside
that reinitiated a new loadOptions function for every render (due to
various reasons such as params used inside changing).

Doing useMemo or useCallback on the search function yielded no suggestion
from the component.

This change introduces a new way of debouncing, using a timeout outside of react
life cycle.

This change also removes the debounceOptions from the search bar component.
Technically it is a breaking change but it is not used anywhere.