fix(framework): support types for "as" component in Button
This commit fixes the issue when specifying "as" property
on the button component would not lead to the base component
inferring props from component specified in "as".

This code would throw an error saying that "layoutId"
is not a valid property:
  <HStack as={motion.div} layoutId="a" />

This commit fixes the issue, by parsing available assume
component props and appending them to the base props (ButtonProps).