fix(framework): fix combopicker to not handle state
Previously combopicker and formatted number input had state handling
that led to stale values being displayed. This commit fixes so that
the values displayed are the latest values that are passed into the
components. The reason for this fix is that handsontable does not
remount the editor between cells. This lead to a stale state persisting
after editing one cell and going to another.

related to dev-21146